Saving settings
Za każdym razem gdy gracz wyjdzie z gry, jego ustawienia się zapisują

+++ include/resources.h
@@ -57,8 +57,8 @@ void initMusicResources();
 
 typedef struct {
     int fps_index;// = 1;
-	float volume; // = 0.5f;
-	float sfx_volume; // = 0.5f;
+	int volume; // = 0.5f;
+	int sfx_volume; // = 0.5f;
 } SettingsComponent;
 void initSettingsComponent();
